show version and release for overlaid components suggestions
